[Feat] 이해도 체크 생성 API 구현

+++ backend/src/main/java/com/example/Piroin/project/domain/curriculum/repository/CurriculumRepository.java
@@ -11,7 +11,7 @@ StudySession(세션) DB 접근 인터페이스
 Q&A 서비스에서 세션 존재 여부 확인 시 사용
 JpaRepository<엔티티 타입, PK 타입> 을 상속하면 findById, save, delete 등 기본 메서드가 자동으로 제공
 */
-public interface CurriculumRepository extends JpaRepository<StudySession, Integer> {
+public interface CurriculumRepository extends JpaRepository<StudySession, Long> {
     List<StudySession> findByStatusOrderBySessionDateAscDayPartAsc(SessionStatus status);
 
     List<StudySession> findByStatusOrderBySessionDateDescDayPartDesc(SessionStatus status);
